Update volume scaling for Launch Control XL
There's some wonky midi value scaling here to make the decibel values
match the printed marks on the Launch Control XL.  I'd imagine there
will be some variance between launch control devices, but this is pretty
close to what the device I have does.

One issue here is that we jump from 0dB up to +2dB, and skip everything
in the +1dB range.  I'll see how important or negligible that issue is
when I start playing with it.