Great folder migration.
Max really didn't like having the Readme in the same directory as the
max project.  Every time I'd open the project, the README.md file would
automatically be moved to a new folder called "other" inside the
project, which was mostly just a pain in the ass in the context of git
and github.